+++ src/main/scala/com/engineplus/star/meta/MetaTableManage.scala
@@ -39,7 +39,7 @@ object MetaTableManage {
       session.execute(
         s"""
            |CREATE KEYSPACE IF NOT EXISTS $database
-           |WITH replication = {'class': 'SimpleStrategy', 'replication_factor': '3'}
+           |WITH replication = {'class': 'SimpleStrategy', 'replication_factor': '1'}
            |AND durable_writes = true;
         """.stripMargin)
     })
